requirements.txt
Flask==2.3.2
python-docx==0.8.11
werkzeug==2.3.4
Flask
Flask-Cors
PyPDF2
python-docx
pandas
